Refactor SqliteConnection commit hook into separate module
